added openid for rack. It is possible to replace the plugins with gems now, since rack-openid is installed.
That means remove the vendored authlogic shit and just use a gem.
Anyway, added rack-openid to the required dependencies and installed authlogic as a plugin (rumored to be better)
pre-prended "/" to all the named routes
Changed some of the path calls in the views.

To get it to work, basically "gem install rack-openid"
And then it should be no biggie.

rake db:drop; rake db:create; rake db:migrate #if you got them rake db:seed
//Seivan